requirements.txt
uvicorn
torch
torchvision
pillow
fastapi
ultralytics
numpy
